requirements.txt
Flask==3.0.0
nsedt==0.0.12
gunicorn>=22.0.0
ta==0.11.0
sklearn==0.0
matplotlib==3.8.2
plotly==5.18.0
gevent==23.9.1
feedparser==6.0.11
newspaper3k==0.2.8
lxml_html_clean==0.2.2
Flask-APScheduler==1.13.1
Flask-SocketIO==5.4.1